pyproject.toml
[project]
name = "towerdefence"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "pygame-ce>=2.5.6",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]


[tool.ruff]
line-length = 120
target-version = "py314"
include = ["src/**/*.py"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint]
ignore = ["COM812"]
select = [
    "E",
    "F",
    "I",
    "A",
    "B",
    "ANN",
    "FAST",
    "ASYNC",
    "ARG",
    "S",
    "BLE",
    "COM",
    "C4",
    "C90",
    "T20",
    "Q",
    "RET",
    "SIM",
    "TC",
    "PTH",
    "N",
    "PERF",
    "W",
    "RUF",
    "ERA",
    "UP",
    "PLC",
    "RET",
]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = [
    "pydantic.BaseModel",
]
